requirements.txt
openai-whisper
torch
torchvision
torchaudio
numpy
setuptools-rust